Restrict proxy to configured server connection pool user
All backend connections in the server connection pool are created
for the same user, which is server_connection_pool_user in the
maxscale config file. For mysql service health check, a second
user service_health_check_user is configured to be able to
connect to Maxscale dbproxy. Other users connections should be
rejected with error. Even maxscale router service user and
monitor user are not allowed to connect to the dbproxy because
they are internally used by Maxscale. For now, other users are
allowed but they would not use pooling connections.